requirements.txt
bs4
unidecode
numpy
pandas
spacy
PyStemmer